backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
redis
requests
